#4 migrated to postgres

+++ pkg/models/character.go
@@ -11,7 +11,7 @@ type Character struct {
 	ID       int           `json:"id"`
 	Name     string        `json:"name"`
 	Type     CharacterType `json:"type"`
-	Scale    float64       `json:"scale"`
+	Scale    float32       `json:"scale"`
 	Position Position      `json:"position"`
 	Image    string        `json:"image"`
 }
